功能列表移至Vuex，新增UserInfo來檢視、編輯會員資訊，另將共用的UserInfoLabel從Register拉出來製作成Component，Alert Component改名成Confirm Component
