--- src/components/auth/Login.vue
@@ -1,14 +1,0 @@
-<template>
-  <div class="login">
-    <h1>Login Page</h1>
-    <form>
-      <label>User Name</label>
-      <input type="text" required />
-      <br />
-      <label>Password</label>
-      <input type="password" required />
-      <br />
-      <button type="submit">Log In</button>
-    </form>
-  </div>
-</template>
